djchat/requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==23.1.0
black==23.3.0
channels==4.0.0
click==8.1.3
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
flake8==6.0.0
h11==0.14.0
httptools==0.5.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
uritemplate==4.1.1
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
